make transfer icons global components
